refactor: 인수 테스트 API 호출을 전부 Steps로 이동
