+++ backend/src/test/java/com/funeat/acceptance/auth/AuthAcceptanceTest.java
@@ -1,12 +1,10 @@
 package com.funeat.acceptance.auth;
 
-import static io.restassured.RestAssured.given;
+import static com.funeat.acceptance.auth.LoginSteps.카카오_로그인_버튼_클릭;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.funeat.acceptance.common.AcceptanceTest;
 import com.funeat.auth.application.AuthService;
-import io.restassured.response.ExtractableResponse;
-import io.restassured.response.Response;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -27,14 +25,5 @@ public class AuthAcceptanceTest extends AcceptanceTest {
 
         // then
         assertThat(actual).isEqualTo(expected);
-    }
-
-    private ExtractableResponse<Response> 카카오_로그인_버튼_클릭() {
-        return given()
-                .redirects().follow(false)
-                .when()
-                .get("/api/auth/kakao")
-                .then()
-                .extract();
     }
 }
